refactor: added use-submit hooks for modals
