+++ shared/fragments/modals/src/modal-create-project/modal-create-project.component.tsx
@@ -1,6 +1,5 @@
 import React                       from 'react'
 import { FC }                      from 'react'
-import { useRouter }               from 'next/router'
 import { useState }                from 'react'
 import { useIntl }                 from 'react-intl'
 
@@ -11,16 +10,14 @@ import { Modal }                   from '@ui/modal'
 
 import { ModalCreateProjectProps } from './modal-create-project.interfaces'
 import { getFormValues }           from '../helpers'
-import { useCreateProject }        from './data'
+import { useSubmit }               from './hooks'
 
 export const ModalCreateProject: FC<ModalCreateProjectProps> = ({ modalOpen = false, onClose }) => {
   const [formValues, setFormValues] = useState<FormProjectValues>(getFormValues())
   const [submitDisabled, setSubmitDisabled] = useState<boolean>(true)
   const { formatMessage } = useIntl()
 
-  const router = useRouter()
-
-  const { createProject } = useCreateProject()
+  const { submit } = useSubmit()
 
   const closeModal = () => {
     setFormValues(getFormValues())
@@ -32,25 +29,8 @@ export const ModalCreateProject: FC<ModalCreateProjectProps> = ({ modalOpen = fa
   }
 
   const handleSubmit = async () => {
-    try {
-      const { avatar, description, name, tags } = formValues
-
-      const avatarFile = await fetch(avatar).then((r) => r.blob())
-
-      const { data } = await createProject({
-        variables: {
-          description,
-          name,
-          tags,
-          avatar: avatarFile,
-        },
-      })
-
-      router.push(`/project/${data?.projectCreate.id}`)
-      closeModal?.()
-    } catch (error) {
-      /** @todo error notification */
-    }
+    await submit(formValues)
+    closeModal?.()
   }
 
   return (
