+++ shared/fragments/modals/src/modal-create-team/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-submit.hook'
