+++ shared/fragments/modals/src/modal-create-team/modal-create-team.component.tsx
@@ -1,6 +1,5 @@
 import React                    from 'react'
 import { FC }                   from 'react'
-import { useRouter }            from 'next/router'
 import { useState }             from 'react'
 import { useIntl }              from 'react-intl'
 
@@ -11,16 +10,14 @@ import { Modal }                from '@ui/modal'
 
 import { ModalCreateTeamProps } from './modal-create-team.interfaces'
 import { getFormValues }        from '../helpers'
-import { useCreateTeam }        from './data'
+import { useSubmit }            from './hooks'
 
 export const ModalCreateTeam: FC<ModalCreateTeamProps> = ({ modalOpen = false, onClose }) => {
   const [formValues, setFormValues] = useState<FormTeamValues>(getFormValues())
   const [submitDisabled, setSubmitDisabled] = useState<boolean>(true)
   const { formatMessage } = useIntl()
 
-  const router = useRouter()
-
-  const { createTeam } = useCreateTeam()
+  const { submit } = useSubmit()
 
   const closeModal = () => {
     setFormValues(getFormValues())
@@ -32,25 +29,8 @@ export const ModalCreateTeam: FC<ModalCreateTeamProps> = ({ modalOpen = false, o
   }
 
   const handleSubmit = async () => {
-    try {
-      const { avatar, description, name, tags } = formValues
-
-      const avatarFile = await fetch(avatar).then((r) => r.blob())
-
-      const { data } = await createTeam({
-        variables: {
-          description,
-          name,
-          tags,
-          avatar: avatarFile,
-        },
-      })
-
-      router.push(`/team/${data?.teamCreate.id}`)
-      closeModal?.()
-    } catch (error) {
-      /** @todo error notification */
-    }
+    await submit(formValues)
+    closeModal?.()
   }
 
   return (
