+++ shared/fragments/modals/src/modal-edit-project/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-submit.hook'
