+++ shared/fragments/modals/src/modal-edit-project/modal-edit-project.component.tsx
@@ -1,17 +1,16 @@
-import React                      from 'react'
-import { FC }                     from 'react'
-import { useState }               from 'react'
-import { useIntl }                from 'react-intl'
+import React                     from 'react'
+import { FC }                    from 'react'
+import { useState }              from 'react'
+import { useIntl }               from 'react-intl'
 
-import { FormProject }            from '@shared/forms-fragment'
-import { FormProjectValues }      from '@shared/forms-fragment'
-import { Layout }                 from '@ui/layout'
-import { Modal }                  from '@ui/modal'
+import { FormProject }           from '@shared/forms-fragment'
+import { FormProjectValues }     from '@shared/forms-fragment'
+import { Layout }                from '@ui/layout'
+import { Modal }                 from '@ui/modal'
 
-import { ModalEditProjectProps }  from './modal-edit-project.interfaces'
-import { getFormValues }          from '../helpers'
-import { useUpdateProject }       from './data'
-import { useUploadProjectAvatar } from './data'
+import { ModalEditProjectProps } from './modal-edit-project.interfaces'
+import { getFormValues }         from '../helpers'
+import { useSubmit }             from './hooks'
 
 export const ModalEditProject: FC<ModalEditProjectProps> = ({
   modalOpen = false,
@@ -23,8 +22,7 @@ export const ModalEditProject: FC<ModalEditProjectProps> = ({
   const [submitDisabled, setSubmitDisabled] = useState<boolean>(true)
   const { formatMessage } = useIntl()
 
-  const { updateProject } = useUpdateProject()
-  const { uploadProjectAvatar } = useUploadProjectAvatar()
+  const { submit } = useSubmit(project)
 
   const closeModal = () => {
     setFormValues(getFormValues(project))
@@ -36,27 +34,11 @@ export const ModalEditProject: FC<ModalEditProjectProps> = ({
   }
 
   const handleSubmit = async () => {
-    try {
-      const { avatar, description, name, tags } = formValues
+    const editedProject = await submit(formValues)
 
-      if (project?.avatarUrl !== avatar) {
-        const avatarFile = await fetch(avatar).then((r) => r.blob())
-
-        await uploadProjectAvatar({ variables: { id: project?.id, file: avatarFile } })
-      }
-
-      await updateProject({ variables: { id: project?.id, name, description, tags } })
-
-      onSubmit?.({
-        ...project,
-        projectName: name,
-        description,
-        tags,
-        avatarUrl: avatar,
-      })
+    if (editedProject) {
+      onSubmit?.(editedProject)
       closeModal?.()
-    } catch (error) {
-      /** @todo error notification */
     }
   }
