+++ shared/fragments/modals/src/modal-edit-team/modal-edit-team.component.tsx
@@ -1,17 +1,16 @@
-import React                   from 'react'
-import { FC }                  from 'react'
-import { useState }            from 'react'
-import { useIntl }             from 'react-intl'
+import React                  from 'react'
+import { FC }                 from 'react'
+import { useState }           from 'react'
+import { useIntl }            from 'react-intl'
 
-import { FormTeam }            from '@shared/forms-fragment'
-import { FormTeamValues }      from '@shared/forms-fragment'
-import { Layout }              from '@ui/layout'
-import { Modal }               from '@ui/modal'
+import { FormTeam }           from '@shared/forms-fragment'
+import { FormTeamValues }     from '@shared/forms-fragment'
+import { Layout }             from '@ui/layout'
+import { Modal }              from '@ui/modal'
 
-import { ModalEditTeamProps }  from './modal-edit-team.interfaces'
-import { getFormValues }       from '../helpers'
-import { useUpdateTeam }       from './data'
-import { useUploadTeamAvatar } from './data'
+import { ModalEditTeamProps } from './modal-edit-team.interfaces'
+import { getFormValues }      from '../helpers'
+import { useSubmit }          from './hooks'
 
 export const ModalEditTeam: FC<ModalEditTeamProps> = ({
   modalOpen = false,
@@ -23,8 +22,7 @@ export const ModalEditTeam: FC<ModalEditTeamProps> = ({
   const [submitDisabled, setSubmitDisabled] = useState<boolean>(true)
   const { formatMessage } = useIntl()
 
-  const { updateTeam } = useUpdateTeam()
-  const { uploadTeamAvatar } = useUploadTeamAvatar()
+  const { submit } = useSubmit(team)
 
   const closeModal = () => {
     setFormValues(getFormValues(team))
@@ -36,27 +34,11 @@ export const ModalEditTeam: FC<ModalEditTeamProps> = ({
   }
 
   const handleSubmit = async () => {
-    try {
-      const { avatar, description, name, tags } = formValues
+    const editedTeam = await submit(formValues)
 
-      if (team?.avatarUrl !== avatar) {
-        const avatarFile = await fetch(avatar).then((r) => r.blob())
-
-        await uploadTeamAvatar({ variables: { id: team?.id, file: avatarFile } })
-      }
-
-      await updateTeam({ variables: { id: team?.id, name, description, tags } })
-
-      onSubmit?.({
-        ...team,
-        name,
-        description,
-        tags,
-        avatarUrl: avatar,
-      })
+    if (editedTeam) {
+      onSubmit?.(editedTeam)
       closeModal?.()
-    } catch (error) {
-      /** @todo error notification */
     }
   }
